Move deinterlacing and RGBA conversions
